refactor: move AppHeader
