+++ src/components/SharedComponents/Layout/Layout.tsx
@@ -5,7 +5,7 @@ import styled from "styled-components";
 
 import Seo, { SeoProps } from "../../../browser-lib/seo/Seo";
 
-import AppHeader from "@/components/AppHeader";
+import AppHeader from "@/components/AppComponents/AppHeader";
 import SiteFooter from "@/components/SiteFooter";
 import { OrganizationJsonLd } from "@/browser-lib/seo/getJsonLd";
 import background, { BackgroundProps } from "@/styles/utils/background";
